refactor: export ProServices, and ValidatorOptions in __init__.py
